Use global `window` object as parent window for elements
I'm running a Vue-based web app inside an iframe in a different domain. In Firefox, Hammer.js generates a lot of random "permission denied" errors from its event handlers when trying to read event type. None of the other components trigger anything remotely similar.

Having researched a lot and fixed the other likely sources of the issue, I've decided that the most likely problem is that Hammer.js uses the wrong `window` to attach its event listeners.

This is an attempt to validate the assumption. I've replaced the current window detection magic with a hard-coded reference to current global object, so that Hammer.js would always use it, and nothing else.